+++ etc/blocks/monetization-test/model/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @milaboratories/milaboratories.monetization-test.model
 
+## 1.2.27
+
+### Patch Changes
+
+- @platforma-sdk/model@1.78.4
+
+## 1.2.26
+
+### Patch Changes
+
+- @platforma-sdk/model@1.78.2
+
+## 1.2.25
+
+### Patch Changes
+
+- Updated dependencies [9650c2b]
+  - @platforma-sdk/model@1.78.1
+
 ## 1.2.24
 
 ### Patch Changes