+++ etc/blocks/pool-explorer/model/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @milaboratories/milaboratories.pool-explorer.model
 
+## 1.3.27
+
+### Patch Changes
+
+- @platforma-sdk/model@1.78.4
+
+## 1.3.26
+
+### Patch Changes
+
+- @platforma-sdk/model@1.78.2
+
+## 1.3.25
+
+### Patch Changes
+
+- Updated dependencies [9650c2b]
+  - @platforma-sdk/model@1.78.1
+
 ## 1.3.24
 
 ### Patch Changes