+++ etc/blocks/pool-explorer/ui/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @milaboratories/milaboratories.pool-explorer.ui
 
+## 1.3.33
+
+### Patch Changes
+
+- Updated dependencies [b0c2b5f]
+  - @platforma-sdk/ui-vue@1.78.4
+  - @milaboratories/milaboratories.pool-explorer.model@1.3.27
+
+## 1.3.32
+
+### Patch Changes
+
+- @platforma-sdk/ui-vue@1.78.2
+- @milaboratories/milaboratories.pool-explorer.model@1.3.26
+
+## 1.3.31
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.model@1.3.25
+- @platforma-sdk/ui-vue@1.78.1
+
 ## 1.3.30
 
 ### Patch Changes