+++ etc/blocks/ui-examples/model/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @milaboratories/milaboratories.ui-examples.model
 
+## 1.6.27
+
+### Patch Changes
+
+- @platforma-sdk/model@1.78.4
+
+## 1.6.26
+
+### Patch Changes
+
+- @platforma-sdk/model@1.78.2
+
+## 1.6.25
+
+### Patch Changes
+
+- Updated dependencies [9650c2b]
+  - @platforma-sdk/model@1.78.1
+
 ## 1.6.24
 
 ### Patch Changes