+++ lib/node/pl-deployments/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @milaboratories/pl-local
 
+## 3.0.4
+
+### Patch Changes
+
+- Updated dependencies [b0c2b5f]
+  - @milaboratories/pl-model-common@1.45.0
+
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies [fbff717]
+  - @milaboratories/pl-model-common@1.44.0
+
 ## 3.0.2
 
 ### Patch Changes