+++ lib/node/pl-drivers/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @milaboratories/pl-drivers
 
+## 1.15.3
+
+### Patch Changes
+
+- Updated dependencies [b0c2b5f]
+  - @milaboratories/pl-model-common@1.45.0
+  - @milaboratories/pl-client@3.11.1
+  - @milaboratories/pl-tree@1.12.9
+
+## 1.15.2
+
+### Patch Changes
+
+- Updated dependencies [e61785b]
+  - @milaboratories/pl-client@3.11.0
+  - @milaboratories/pl-tree@1.12.8
+
+## 1.15.1
+
+### Patch Changes
+
+- Updated dependencies [fbff717]
+  - @milaboratories/pl-model-common@1.44.0
+  - @milaboratories/pl-client@3.10.2
+  - @milaboratories/pl-tree@1.12.7
+
 ## 1.15.0
 
 ### Minor Changes