+++ lib/ptabler/schema/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @platforma-open/software-ptabler.types
 
+## 1.15.12
+
+### Patch Changes
+
+- Updated dependencies [b0c2b5f]
+  - @milaboratories/pl-model-common@1.45.0
+
+## 1.15.11
+
+### Patch Changes
+
+- Updated dependencies [fbff717]
+  - @milaboratories/pl-model-common@1.44.0
+
 ## 1.15.10
 
 ### Patch Changes