+++ tools/tengo-builder/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @platforma-sdk/tengo-builder
 
+## 4.0.5
+
+### Patch Changes
+
+- @milaboratories/pl-model-backend@1.4.4
+
+## 4.0.4
+
+### Patch Changes
+
+- @milaboratories/pl-model-backend@1.4.3
+
+## 4.0.3
+
+### Patch Changes
+
+- @milaboratories/pl-model-backend@1.4.2
+
 ## 4.0.2
 
 ### Patch Changes